Thie PR adds the FastOpt object and extends the Options object by
* Using YAML only as a background option container that is used for lazily updating the FastOpt object.
* The FastOpt object gets rebuilt whenever an access is about to occur and the doLazyRebuild_ flag is set.
* This will be the case whenever the underlying YAML object has been modified through the public interface.
* The YAML object itself it not accessible by the outside anymore (or only through cloning) to avoid untrackable modifications.

This seems to result in a >20% speed-up for decoding.

+++ src/models/bert.h
@@ -283,7 +283,7 @@ public:
                     .push_back(mlp::dense()                           //
                                  ("prefix", prefix_ + "_ff_logit_l1") //
                                  ("dim", dimModel)                    //
-                                 ("activation", mlp::act::tanh))      // @TODO: do we actually need this?
+                                 ("activation", (int)mlp::act::tanh))      // @TODO: do we actually need this?
                     .push_back(mlp::output()                          //
                                  ("dim", dimTrgCls))                  //
                                  ("prefix", prefix_ + "_ff_logit_l2") //
